+++ output/idcsv/idcsv.go
@@ -23,7 +23,14 @@ var total uint
 var ids = make([]string, 0)
 
 func init() {
-	output.Register("idcsv", setup, start, process, finish)
+	output.Register("idcsv", supports, setup, start, process, finish)
+}
+
+func supports(outtype string) bool {
+	if outtype == "probe" || outtype == "anchor" || outtype == "msm" {
+		return true
+	}
+	return false
 }
 
 func setup(isverbose bool, options []string) {