+++ output/most/most.go
@@ -15,6 +15,7 @@ import (
 	"goatcli/output"
 	"goatcli/output/annotate"
 	"goatcli/output/some"
+	"slices"
 	"strings"
 
 	"github.com/robert-kisteleki/goatapi"
@@ -25,7 +26,17 @@ var verbose bool
 var total uint
 
 func init() {
-	output.Register("most", setup, start, process, finish)
+	output.Register("most", supports, setup, start, process, finish)
+}
+
+func supports(outtype string) bool {
+	if slices.Contains([]string{"ping", "trace", "dns", "tls", "ntp", "http"}, outtype) ||
+		outtype == "connection" || outtype == "uptime" ||
+		outtype == "probe" || outtype == "anchor" || outtype == "msm" ||
+		outtype == "status" {
+		return true
+	}
+	return false
 }
 
 func setup(isverbose bool, options []string) {