+++ output/none/none.go
@@ -13,13 +13,24 @@ package none
 import (
 	"fmt"
 	"goatcli/output"
+	"slices"
 )
 
 var verbose bool
 var total uint
 
 func init() {
-	output.Register("none", setup, start, process, finish)
+	output.Register("none", supports, setup, start, process, finish)
+}
+
+func supports(outtype string) bool {
+	if slices.Contains([]string{"ping", "trace", "dns", "tls", "ntp", "http"}, outtype) ||
+		outtype == "connection" || outtype == "uptime" ||
+		outtype == "probe" || outtype == "anchor" || outtype == "msm" ||
+		outtype == "status" {
+		return true
+	}
+	return false
 }
 
 func setup(isverbose bool, options []string) {