+++ doc/changes.md
@@ -5,7 +5,7 @@
 * FIX: native trace out didn't show hop errors nicely
 * NEW: add shortcuts for measurement types, e.g. `goatcli ping ping.me` just works
 * NEW: pre-flight checks for output format
-* NEW: add `type:TYPE` option to the dnsstat output formatter
+* NEW: add `type:TYPE` option to the dnsstat output formatter, automatically add this for new DNS measurements
 * CHANGED: update to go1.21 and update dependencies
 
 ## v0.5.0